requirements.txt
discord.py
python-decouple
PyNaCl
youtube-dl